+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [1.9.1] - 2020-04-03
+### Added
+- configuration of REST clients without having to use the API
+- support for learning modules (htlm, sahs)
+
 ## [1.9.0] - 2020-04-02
 ### Added
 - eBook cover route